main, GitCommitReader: Don't use --stat with partial clones
The --stat (or --shortstat) parameter would cause us to download lots
of data from origin for relatively little benefit. Walking the full
repository this is also fairly inefficient.

We warn when this happens so the user knows not to rely on the change
statistic, or to make a full clone if needed.

Discussed in GitHub #13.